Refactor: 경로 오류 수정
