+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import GlobalStyles from '@/styles/globalStyles';
+import GlobalStyles from '@/styles/GlobalStyles';
 import type { AppProps } from 'next/app'
 
 export default function App({ Component, pageProps }: AppProps) {
