+++ styles/GlobalStyles.tsx
